requirements.txt
pandas
apache-airflow
apache-airflow-providers-postgres